+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix missing labels in [`diff`] output. [#1026]
+
 ## [1.9.0] - 2022-06-16
 
 ### Added